feat: Add giveaway editing functionality
The update includes changes to the `discord` package, which adds the `do-proxy` package as a dependency and updates `slash-create` to `5.12.0`. Also, new files were added to the project, such as `editModal.ts`. Additionally, changes to `index.ts` and `edit.ts` help provide functionality for editing giveaways. Finally, the `CFWorkerServer` class gained a new `states` property, which allows durable objects to be accessed.